+++ app/app.py
@@ -2,7 +2,7 @@ from flask import Flask, jsonify, request
 
 from data_sources.datastore import get_call_history_records_by_interviewer_and_date_range, \
     get_call_history_report_status
-from interviewer_pattern_report.report import get_call_pattern_records_by_interviewer_and_date_range
+from interviewer_call_pattern_report.report import get_call_pattern_records_by_interviewer_and_date_range
 from models.config import Config
 
 app = Flask(__name__)
