+++ functions/google_storage_functions.py
